third_party/blink/web_tests/fast/dom/inline-event-attributes-lookup-expected.txt

Tests the lookup in inline event handlers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is "test-input-a"
PASS result is "test-form-b"
PASS result is "document-c"
PASS result is "div-a"
PASS result is "document-b"
PASS result is "document-c"
PASS result is div
PASS result is not "FAIL"
PASS successfullyParsed is true

TEST COMPLETE
